Put it a dedicated worker
