+++ front/start_worker.ts
@@ -4,6 +4,7 @@ import { hideBin } from "yargs/helpers";
 
 import logger from "@app/logger/logger";
 import { runPokeWorker } from "@app/poke/temporal/worker";
+import { runDataRetentionWorker } from "@app/temporal/data_retention/worker";
 import { runHardDeleteWorker } from "@app/temporal/hard_delete/worker";
 import { runLabsWorker } from "@app/temporal/labs/worker";
 import { runMentionsCountWorker } from "@app/temporal/mentions_count_queue/worker";
@@ -32,7 +33,8 @@ type WorkerName =
   | "scrub_workspace_queue"
   | "update_workspace_usage"
   | "upsert_queue"
-  | "upsert_table_queue";
+  | "upsert_table_queue"
+  | "data_retention";
 
 const workerFunctions: Record<WorkerName, () => Promise<void>> = {
   hard_delete: runHardDeleteWorker,
@@ -47,6 +49,7 @@ const workerFunctions: Record<WorkerName, () => Promise<void>> = {
   update_workspace_usage: runUpdateWorkspaceUsageWorker,
   upsert_queue: runUpsertQueueWorker,
   upsert_table_queue: runUpsertTableQueueWorker,
+  data_retention: runDataRetentionWorker,
 };
 const ALL_WORKERS = Object.keys(workerFunctions);
