+++ front/temporal/data_retention/config.ts
@@ -1,0 +1,3 @@
+const QUEUE_VERSION = 1;
+
+export const QUEUE_NAME = `data-retention-queue-v${QUEUE_VERSION}`;
