+++ front/temporal/data_retention/utils.ts
@@ -1,0 +1,7 @@
+/**
+ * Purge Data retention logic.
+ */
+
+export function getPurgeDataRetentionScheduleId() {
+  return "purge-data-retention-schedule";
+}
