+++ front/temporal/hard_delete/utils.ts
@@ -22,11 +22,3 @@ export function getRunExecutionsDeletionCutoffDate(): number {
 
   return cutoffDate.getTime();
 }
-
-/**
- * Purge Data retention logic.
- */
-
-export function getPurgeDataRetentionScheduleId() {
-  return "purge-data-retention-schedule";
-}
