+++ front/temporal/hard_delete/workflows.ts
@@ -1,5 +1,4 @@
 import { proxyActivities } from "@temporalio/workflow";
-import _ from "lodash";
 
 import type * as activities from "@app/temporal/hard_delete/activities";
 
@@ -9,24 +8,7 @@ const { purgeExpiredRunExecutionsActivity } = proxyActivities<
 >({
   startToCloseTimeout: "60 minutes",
 });
-const {
-  getWorkspacesWithConversationsRetentionActivity,
-  purgeConversationsBatchActivity,
-} = proxyActivities<typeof activities>({
-  startToCloseTimeout: "15 minutes",
-});
 
 export async function purgeRunExecutionsCronWorkflow(): Promise<void> {
   await purgeExpiredRunExecutionsActivity();
-}
-
-export async function purgeDataRetentionWorkflow(): Promise<void> {
-  const workspaceIds = await getWorkspacesWithConversationsRetentionActivity();
-  const workspaceChunks = _.chunk(workspaceIds, 4);
-
-  for (const workspaceChunk of workspaceChunks) {
-    await purgeConversationsBatchActivity({
-      workspaceIds: workspaceChunk,
-    });
-  }
 }
